refactor: rename response to content
