+++ app/sol/agent/[address]/chat/content/issue-token/constants.ts
