+++ app/sol/agent/[address]/chat/content/trade.tsx/buy.tsx
