+++ app/sol/agent/[address]/chat/content/trade.tsx/sell.tsx
