+++ app/sol/agent/[address]/chat/content/types.ts
